requirements.txt
dj-database-url==0.5.0
dj-email-url==1.0.5
django-cache-url==3.3.0
Django==4.0.2
django-cors-headers==3.8.0
djangorestframework==3.12.4
PyJWT==2.3.0
pytest==6.2.5
pytest-django==4.5.1
pytest-xdist==2.5.0
pytz==2022.1
